Search for the second biggest integer in the list of arguments

+++ 0x12-javascript-warm_up/11-second_biggest.js
@@ -11,7 +11,7 @@ function createArray () {
   let i = 2;
 
   while (i < len) {
-    myArray.push(parseInt(process.argv[i]));
+    myArray.push(parseFloat(process.argv[i]));
     i++;
   }
